+++ packages/passkeys/package.json
@@ -1,27 +1,31 @@
 {
-  "name": "passkeys",
-  "version": "1.0.0",
-  "description": "",
-  "type": "module",
-  "main": "index.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@alchemy/aa-accounts": "^*",
-    "@alchemy/aa-alchemy": "^*",
-    "@alchemy/aa-core": "^*",
-    "@alchemy/aa-signers": "^*",
-    "@turnkey/http": "^2.6.2",
-    "@turnkey/viem": "^0.4.7",
-    "viem": "^1.21.4"
-  },
-  "devDependencies": {
-    "@repo/eslint-config": "*",
-    "@repo/typescript-config": "*",
-    "typescript": "^5.3.3",
-    "axios": "^1.6.5"
-  }
+    "name": "@repo/passkeys",
+    "version": "1.0.0",
+    "description": "",
+    "type": "module",
+    "main": "index.ts",
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "exports": {
+        ".": "./src/index.ts",
+        "./model": "./src/model/index.ts"
+    },
+    "author": "",
+    "license": "ISC",
+    "dependencies": {
+        "@alchemy/aa-accounts": "^*",
+        "@alchemy/aa-alchemy": "^*",
+        "@alchemy/aa-core": "^*",
+        "@alchemy/aa-signers": "^*",
+        "@turnkey/http": "^2.6.2",
+        "@turnkey/viem": "^0.4.7",
+        "viem": "^1.21.4"
+    },
+    "devDependencies": {
+        "@repo/eslint-config": "*",
+        "@repo/typescript-config": "*",
+        "typescript": "^5.3.3",
+        "axios": "^1.6.5"
+    }
 }
