+++ packages/passkeys/src/util/encoding.ts
